rag-chatbot-backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
openai==1.12.0
qdrant-client==1.8.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
tiktoken==0.5.2
aiofiles==23.2.1
requests==2.31.0
python-multipart==0.0.9